+++ tybotRoute/CHANGELOG.md
@@ -5,8 +5,12 @@
 available on:
  ▶️ https://www.npmjs.com/package/@tiledesk/tiledesk-tybot-connector
 
+### 0.1.61
+- Added DirRandomReply, TiledeskChatbotUtil.chooseRandomReply()
+ 
 ### 0.1.60
 - fixed: get department_id from message.request instead of message.attributes
+- added JSONOperationToExpression evaluator in TiledeskExpression
 
 ### 0.1.59
 - Log clean